+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Ignore errors when importing activities from outbox.
 - Make activity limit in outbox fetcher adjustable.
 
+### Fixed
+
+- Make `/api/v1/accounts/{account_id}/follow` work with form-data.
+
 ## [1.21.0] - 2023-04-12
 
 ### Added